ok. Binary Trees exist. I want to try that here
I wanna see if I can use Binary Trees for the pages in the array. (or some basic version of it). Was talking to a friend about how to approach the assignment and gave me this idea.
It might be way too advanced for me but I think it would be something worth trying and learning.